Fix summary list action alignment
Fixes the alignment on the Summary List component action links, if there are enough links to wrap onto multiple lines. It does this by reversing the side of the link that margin, padding and borders are applied to depending on the breakpoint.

On mobile, links are left aligned with spacing and borders on the right. On tablet and above, links are right aligned with spacing and borders on the left.

Additionally, I've made this into a flexbox container to remove unevent spacing that was being caused by errant whitespace. On IE10 and below, this falls back to inline-block, and the errant
spaces remain. This also includes a fix for a vertical alignment issue caused by a flexbox container being within a table-cell on IE11.